competitor_monitor/requirements.txt
anthropic==0.30.0
requests==2.31.0
feedparser==6.0.10
python-dotenv==1.0.0
